+++ official/vision/image_classification/pruning/mnist/block_sparsity_example.yaml
@@ -1,0 +1,68 @@
+# Fine-grained pruning configuration for the MNIST model.
+model_name: 'mnist'
+pruning:
+  - layer_name: 'conv2d'
+    pruning:
+      - weight_name: 'kernel'
+        pruning:
+          pruning_schedule:
+            class_name: 'ConstantSparsity'
+            config:
+              target_sparsity: 0.5
+              begin_step: 0
+              end_step: -1
+              frequency: 100
+          pruning_granularity:
+            class_name: 'BlockSparsity'
+            config:
+              block_size: [1, 1]
+              block_pooling_type: 'AVG'
+  - layer_name: 'conv2d_1'
+    pruning:
+      - weight_name: 'kernel'
+        pruning:
+          pruning_schedule:
+            class_name: 'ConstantSparsity'
+            config:
+              target_sparsity: 0.5
+              begin_step: 0
+              end_step: -1
+              frequency: 100
+          pruning_granularity:
+            class_name: 'BlockSparsity'
+            config:
+              block_size: [1, 1]
+              block_pooling_type: 'AVG'
+  - layer_name: 'dense'
+    pruning:
+      - weight_name: 'kernel'
+        pruning:
+          pruning_schedule:
+            class_name: 'ConstantSparsity'
+            config:
+              target_sparsity: 0.5
+              begin_step: 0
+              end_step: -1
+              frequency: 100
+          pruning_granularity:
+            class_name: 'BlockSparsity'
+            config:
+              block_size: [1, 1]
+              block_pooling_type: 'AVG'
+  - layer_name: 'dense_1'
+    pruning:
+      - weight_name: 'kernel'
+        pruning:
+          pruning_schedule:
+            class_name: 'ConstantSparsity'
+            config:
+              target_sparsity: 0.5
+              begin_step: 0
+              end_step: -1
+              frequency: 100
+          pruning_granularity:
+            class_name: 'BlockSparsity'
+            config:
+              block_size: [1, 1]
+              block_pooling_type: 'AVG'
+share_mask: null
