+++ official/vision/image_classification/pruning/pruning_base_configs.py
